pdf being produced wih debug on

+++ t/2-render-core-font.t
@@ -3,6 +3,9 @@ use Test;
 use PDF::API6;
 use App::FontSample;
 
+my $debug = 1;
+note "\$debug is on" if $debug;
+
 my $pdf = PDF::API6.new;
 my $font = $pdf.core-font: :family<Helvetica>;
 my $dir = $*TMPDIR.add("app-fontsample-$*PID");
@@ -15,12 +18,19 @@ my $path = create-font-sample(
     :output($output),
 );
 
+if $debug {
+    say "DEBUG: See PDF file '$path'";
+}
+
 ok $path.f, 'sample PDF was created';
 ok $path.s > 0, 'sample PDF is not empty';
 
 LEAVE {
-    $output.unlink if $output.e;
-    $dir.rmdir if $dir.d;
+    unless $debug {
+        $output.unlink if $output.e;
+        $dir.rmdir if $dir.d;
+    }
 }
+
 
 done-testing;
